:truck: rename: ScrollToTop utils 폴더로 이동 #147
